Temp commenting-out of custom writers
User has many Positions, and Position has many Users; the relationship is defined in the schema through the PositionsUser table.  Rails automatically writes to the PositionsUser table when a user updates their collection of positions.  This custom writer code isn't currently necessary as-written.  Commenting instead of removing because of plans to add a new attribute to PositionsUser, for skill level.